refactor(jolt-kernels): move the reference kernels under reference/
The crate root is now the seam alone: per-relation modules define the
slot's object-safe factory/instance traits, backend.rs holds the
JoltBackend registry and ProofSession, and the wire/grid types stay with
their traits. Every kernel implementation — the ReferenceBackend impls,
the kernel structs, the naive interpreter, the view helpers — moves
verbatim to the module's sibling under reference/, whose mod.rs owns the
ReferenceBackend marker and the JoltBackend::reference() constructor.

The seam never depends on reference/: a future optimized backend is a
sibling directory implementing the same crate-root traits. Crate-root
re-exports (ReferenceBackend, NaiveSumcheckProver) keep downstream paths
unchanged; no stage recipe names a concrete kernel type, so jolt-prover
is untouched.
