+++ crates/jolt-kernels/src/registers_val_evaluation.rs
@@ -1,28 +1,13 @@
-//! The registers value-evaluation (stage 5) kernel: a naive member over the
-//! cycle domain.
-//!
-//! The summand is `LT(j, r_cycle) · rd_inc(j) · rd_wa(r_address, j)` — the
-//! "register value at `(r_address, r_cycle)` is the sum of earlier
-//! increments" identity. The `rd_wa` opening is the address-bound slice of
-//! the `(2^7 × T)` one-hot write-address grid (an opening-side fold), and
-//! `LtCycle` is ONE multilinear: `LtPolynomial::evaluations(r_cycle)`.
+//! The registers value-evaluation (stage 5) slot.
 
-use std::collections::BTreeMap;
-
-use jolt_claims::protocols::jolt::geometry::dimensions::REGISTER_ADDRESS_BITS;
-use jolt_claims::protocols::jolt::geometry::registers::{
-    rd_inc_val_evaluation, rd_wa_val_evaluation,
-};
-use jolt_claims::protocols::jolt::{JoltDerivedId, RegistersValEvaluationPublic, TraceDimensions};
+use jolt_claims::protocols::jolt::TraceDimensions;
 use jolt_claims::NoChallenges;
 use jolt_field::Field;
-use jolt_poly::{BindingOrder, LtPolynomial, Polynomial};
 use jolt_verifier::stages::stage5::registers_val_evaluation::RegistersValEvaluation;
 use jolt_witness::protocols::jolt_vm::JoltVmNamespace;
 use jolt_witness::WitnessProvider;
 
-use crate::views::{dense_view, eq_table};
-use crate::{KernelError, NaiveSumcheckProver, ProofSession, ProveSumcheck, ReferenceBackend};
+use crate::{KernelError, ProofSession, ProveSumcheck};
 
 /// The stage-5 registers value-evaluation slot.
 pub trait RegistersValEvaluationProver<F: Field> {
@@ -34,57 +19,4 @@ pub trait RegistersValEvaluationProver<F: Field> {
         challenges: &NoChallenges<F>,
         witness: &dyn WitnessProvider<F, JoltVmNamespace>,
     ) -> Result<Box<dyn ProveSumcheck<F, Relation = RegistersValEvaluation<F>>>, KernelError<F>>;
-}
-
-impl<F: Field> RegistersValEvaluationProver<F> for ReferenceBackend {
-    fn prepare(
-        &self,
-        _session: &mut ProofSession,
-        trace_dimensions: TraceDimensions,
-        registers_val_point: &[F],
-        challenges: &NoChallenges<F>,
-        witness: &dyn WitnessProvider<F, JoltVmNamespace>,
-    ) -> Result<Box<dyn ProveSumcheck<F, Relation = RegistersValEvaluation<F>>>, KernelError<F>>
-    {
-        let cycles = 1usize << trace_dimensions.log_t();
-        let registers = 1usize << REGISTER_ADDRESS_BITS;
-        if registers_val_point.len() != REGISTER_ADDRESS_BITS + trace_dimensions.log_t() {
-            return Err(KernelError::Unsupported {
-                reason: "registers value-evaluation input point has the wrong variable count",
-            });
-        }
-        let (r_address, r_cycle) = registers_val_point.split_at(REGISTER_ADDRESS_BITS);
-
-        // The address-bound `rd_wa` slice, folded from the one-hot grid.
-        let eq_address = eq_table(r_address);
-        let rd_wa_full = dense_view(witness, rd_wa_val_evaluation())?;
-        let wa_folded: Vec<F> = (0..cycles)
-            .map(|j| {
-                (0..registers)
-                    .map(|k| rd_wa_full[(k << trace_dimensions.log_t()) | j] * eq_address[k])
-                    .sum()
-            })
-            .collect();
-
-        let opening_tables = BTreeMap::from([
-            (rd_wa_val_evaluation(), Polynomial::new(wa_folded)),
-            (
-                rd_inc_val_evaluation(),
-                Polynomial::new(dense_view(witness, rd_inc_val_evaluation())?),
-            ),
-        ]);
-        let derived_tables = BTreeMap::from([(
-            JoltDerivedId::from(RegistersValEvaluationPublic::LtCycle),
-            Polynomial::new(LtPolynomial::evaluations(r_cycle)),
-        )]);
-
-        let relation = RegistersValEvaluation::new(trace_dimensions);
-        Ok(Box::new(NaiveSumcheckProver::new(
-            relation,
-            challenges,
-            opening_tables,
-            derived_tables,
-            BindingOrder::LowToHigh,
-        )?))
-    }
 }
